+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/repository/IBaseChannelRepository.java
@@ -40,6 +40,6 @@ public interface IBaseChannelRepository {
      *
      * @return
      */
-    Observable<BaseJson<List<DynamicBean>>> getDynamicListFromChannel(long channel_id,int max_id);
+    Observable<BaseJson<List<DynamicBean>>> getDynamicListFromChannel(long channel_id,long max_id);
 
 }
